Fix forcing rM1 model number

+++ shared/libblight/socket.cpp
@@ -60,7 +60,7 @@ std::optional<Blight::data_t> Blight::recv(
     }
     // The data we recieved isn't the same size as what we expected
     if(res != size){
-        _WARN("recv %ld != %ld", (long int)size, (long int)res);
+        _WARN("recv %d != %d", size, res);
         delete[] data;
         errno = EBADMSG;
         return {};
@@ -87,7 +87,7 @@ std::optional<Blight::data_t> Blight::recv_blocking(int fd, ssize_t size){
     }
     // The data we recieved isn't the same size as what we expected
     if(res != size){
-        _WARN("recv_blocking %ld != %ld", (long int)size, (long int)res);
+        _WARN("recv_blocking %d != %d", size, res);
         delete[] data;
         errno = EBADMSG;
         return {};
@@ -115,7 +115,7 @@ bool Blight::send_blocking(int fd, const data_t data, ssize_t size){
     }
     // The data we sent isn't the same size as what we expected
     if(res != size){
-        _WARN("send_blocking %ld != %d", (long int)size, res);
+        _WARN("send_blocking %d != %d", size, res);
         errno = EMSGSIZE;
         return false;
     }
